risc-v example for QEMU virt based on picolibc
Add example for QEMU RISC-V virt machine to experiment with gdb scripts.
This example is supposed to run in QEMU on 'virt' RV32 machine. Example
is based on Picolibc using its crt0 startup files and semihosting for
printing. As a result, example app can be written only in C.

Toolchain for this example can be built using crosstool-ng, sample
configuration is provided in 'support' directory.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>